refactor: 不要なコメントアウトの削除、UpdateItemのファイル切り出し
